# web/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rust Tunnel API",
    "description": "RSTunnel 管理面 REST API（§21）",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/routes/{id}": {
      "get": {
        "tags": [
          "routes"
        ],
        "operationId": "get_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Route id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            }
          },
          "404": {
            "description": "Route not found"
          }
        }
      },
      "delete": {
        "tags": [
          "routes"
        ],
        "operationId": "delete_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Route id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Route deleted"
          },
          "404": {
            "description": "Route not found"
          }
        }
      },
      "patch": {
        "tags": [
          "routes"
        ],
        "operationId": "update_route",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Route id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Route updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            }
          },
          "404": {
            "description": "Route not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Limits": {
        "type": "object",
        "properties": {
          "max_bandwidth": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "minimum": 0
          },
          "max_connection_rate": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "minimum": 0
          },
          "max_connections": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "minimum": 0
          }
        }
      },
...
      "Route": {
        "type": "object",
        "description": "Route 响应对象（api.md §6）。",
        "required": [
          "id",
          "name",
          "node_id",
          "type",
          "enabled",
          "target_host",
          "target_port",
          "status"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "hostname": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "limits": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Limits"
              }
            ],
            "nullable": true
          },
          "listen_host": {
            "type": "string",
            "nullable": true
          },
          "listen_port": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "node_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "target_host": {
            "type": "string"
          },
          "target_port": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "tls_mode": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/RouteType"
          }
        }
      },
      "RouteType": {
        "type": "string",
        "enum": [
          "tcp",
          "udp",
          "http",
          "https"
        ]
      },
      "TlsMode": {
        "type": "string",
        "enum": [
          "terminate",
          "passthrough",
          "disabled"
        ]
      },
...
      "UpdateRouteRequest": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "nullable": true
          },
          "hostname": {
            "type": "string",
            "nullable": true
          },
          "limits": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Limits"
              }
            ],
            "nullable": true
          },
          "listen_host": {
            "type": "string",
            "nullable": true
          },
          "listen_port": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "node_id": {
            "type": "string",
            "nullable": true
          },
          "target_host": {
            "type": "string",
            "nullable": true
          },
          "target_port": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "minimum": 0
          },
          "tls_mode": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TlsMode"
              }
            ],
            "nullable": true
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RouteType"
              }
            ],
            "nullable": true
          }
        }
      },
...
    }
  },
...
}